requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==4.0.1
Sphinx
django-extensions==1.5.7
Werkzeug==0.10.4
django-test-plus==1.0.9
factory-boy==2.9.2

# django-debug-toolbar that works with Django 1.5+
django-debug-toolbar==1.4

# improved REPL
ipdb==0.8.1

# Required by maildump. Need to pin dependency to gevent beta to be Python 3-compatible.
gevent==1.0.2
# Enables better email testing
maildump==0.5.1
